add traits.rs file
